+++ Server/Giant.Battle/Component/Scene/Map/MapScene_NPC.cs
@@ -5,8 +5,8 @@ namespace Giant.Battle
 {
     public partial class MapScene
     {
-        private Dictionary<int, NPC> npcList = new Dictionary<int, NPC>();
-        public Dictionary<int, NPC> NpcList => npcList;
+        private Dictionary<long, NPC> npcList = new Dictionary<long, NPC>();
+        public Dictionary<long, NPC> NpcList => npcList;
 
         protected virtual void UpdateNpc(double dt)
         {
